+++ Fogger/src/main/java/dotblueshoes/fogger/Fogger.java
@@ -13,15 +13,10 @@ import org.apache.logging.log4j.Logger;
 
 import dotblueshoes.fogger.ConfigHandler;
 
-@Mod(
-    modid = Fogger.MODID, 
-    version = Fogger.VERSION,
-    useMetadata = true
-)
-
+@Mod( modid = Fogger.MODID,  version = Fogger.VERSION, useMetadata = true)
 public class Fogger {
     public static final String MODID = "fogger";
-    public static final String VERSION = "1.0";
+    public static final String VERSION = "1.1";
 
     @Instance(MODID)
     public static Fogger instance;
@@ -30,13 +25,14 @@ public class Fogger {
 
     @EventHandler
     public void preInit(FMLPreInitializationEvent event) {
-        ConfigHandler.loadConfig(event.getSuggestedConfigurationFile());
+        ConfigHandler.loadConfigurationFile(event.getSuggestedConfigurationFile());
         logger = event.getModLog();
     }
 
     @EventHandler
     public void init(FMLInitializationEvent event) {
         //FogEvent.setupFogRendering();
+        ConfigHandler.biomeFogs[3].yLevel = 60;
         if(ConfigHandler.globalFog)
             MinecraftForge.EVENT_BUS.register(new GlobalFogEvent());
         else